+++ ggFrames/localization/FR.lua
@@ -12,13 +12,16 @@ GGF.locale["Display_Player_Experience"] = "Afficher la barre d'expérience"
 
 -- Global
 GGF.locale['Combat_Alpha']              = "Hors de combat Alpha"
-GGF.locale['Font']                      = "Font" -- I would leave it that way
+GGF.locale['Font']                      = "Font"
 GGF.locale['Font_Style']                = "Style du font"
 GGF.locale['Bar_Texture']               = "Texture"
 GGF.locale["Bar_Format"]                = "Format de la barre d'affichage de texte"
 GGF.locale["Bar_Format_Location"]       = "La position de la barre d'affichage de texte"
 GGF.locale["Bar_Format_Two"]            = "Deuxième format de la barre d'affichage de texte"
 GGF.locale["Bar_Format_Location_Two"]   = "La position de la deuxième barre d'affichage"
+GGF.locale["Shield_Height"]             = "Bouclier Hauteur"                                                            -- Need Proper Translation
+GGF.locale["Shield_BarColor"]           = "Couleur Bouclier"                                                            -- Need Proper Translation
+GGF.locale["Mount_Seperated"]           = "Séparée Cadre Mont"                                                          -- Need Proper Translation
 
 -- Frame Options
 GGF.locale["Frame_PositionX"]           = "La position du cadre horizontal"
@@ -60,9 +63,9 @@ GGF.locale['Setting_Player_Stamina']    = "Cadre du joueur: Points d'endurance"
 GGF.locale['Setting_Player_Experience'] = "Cadre du jouer: Expérience"
 GGF.locale['Setting_Player_Mount']      = "Cadre du jouer: Monture"
 GGF.locale['Setting_Group']             = "Cadre de groupe"
-GGF.locale['Setting_Raid']              = "Cadre de raid" -- I would leave it to raid
+GGF.locale['Setting_Raid']              = "Cadre de raid"
 GGF.locale['Setting_Raid_Layout']       = "L'affichage du cadre de raid"
-GGF.locale['Setting_Raid_Layout_Opt']  = {'Une rangée de six groupes', 'Deux rangées de trois groupes'}
+GGF.locale['Setting_Raid_Layout_Opt']   = {'Une rangée de six groupes', 'Deux rangées de trois groupes'}                -- Need Proper Translation
 GGF.locale['Setting_Target']            = "Cadre de la cible"
 GGF.locale['Setting_Defaults']          = "Remettre les configurations par défaut"
 GGF.locale['Setting_Reset_Warning']     = "Attention: En remettant les configurations par défaut, vous allez perdre toutes configurations manuelle. Assurez-vous que c'est bel et bien ce que vous voulez avant de presser le bouton du bas."