+++ ggFrames/unit/Functions.lua
@@ -19,7 +19,7 @@ function GGF.Unit:Load(unitTag)
   self:SetDeath( IsUnitDead(self.unitTag) )
   self:SetLevel( GetUnitLevel(self.unitTag), GetUnitVeteranRank(self.unitTag) )             -- Level / Experience
   self:SetRange( IsUnitInGroupSupportRange(self.unitTag) )                                  -- Is Within Support Range
-  
+
   if self.unitName == "Target" then
     self:SetCaption( GetUnitCaption(self.unitTag) or GetUnitTitle(self.unitTag) )
     self:SetAllianceRank( GetUnitAvARank(self.unitTag) )
@@ -27,7 +27,7 @@ function GGF.Unit:Load(unitTag)
   end
   
   if IsUnitPlayer(self.unitTag) then
-    self:SetClass( GetUnitClass(self.unitTag) )                                             -- Class Texture
+    self:SetClass( GetUnitClassId(self.unitTag) )                                             -- Class Texture
     self:SetLeader( IsUnitGroupLeader(self.unitTag) )                                       -- Is Group Leader (Non-Players are never Group Leaders)
   else 
     self:SetClass( nil )
@@ -35,6 +35,7 @@ function GGF.Unit:Load(unitTag)
   end
 
   self:SetPower( nil, POWERTYPE_HEALTH, GetUnitPower(self.unitTag, POWERTYPE_HEALTH) )      -- Set Health
+  self:UpdateShield( GetUnitAttributeVisualizerEffectInfo(self.unitTag, ATTRIBUTE_VISUAL_POWER_SHIELDING, STAT_MITIGATION, ATTRIBUTE_HEALTH, POWERTYPE_HEALTH) )
   if self.unitName == "Player" then
     self:SetPower( nil, POWERTYPE_MAGICKA, GetUnitPower(self.unitTag, POWERTYPE_MAGICKA) )  -- Set Magicka
     self:SetPower( nil, POWERTYPE_STAMINA, GetUnitPower(self.unitTag, POWERTYPE_STAMINA) )  -- Set Stamina
@@ -80,13 +81,25 @@ end
 function GGF.Unit:SetMountPower( powerIndex, powerType, powerValue, powerMax, powerEffectiveMax )
   if not self.template.Mount then return end
   self.mount = {current = powerValue, max = powerEffectiveMax, percent = ( ( powerValue / powerEffectiveMax ) * 100 )}
-  self.frames.mountSt:SetWidth( ( self.mount.percent / 100 ) * self.template.Mount.BarArea.Bar.Width )
+  local barWidth = GGF.SavedVars['Player_Mount_Seperated'] and self.template.Mount.BarArea.Bar.Width or self.template.Mount.Bar.Width
+  self.frames.mountSt:SetWidth( ( self.mount.percent / 100 ) * barWidth )
 end
 
 function GGF.Unit:SetMounted( isMounted )
   if not self.template.Mount then return end
   self.isMounted = isMounted
   self.frames.mount:SetHidden( not self.isMounted )
+  if not GGF.SavedVars['Player_Mount_Seperated'] then
+    GGF.Window:SetOffsetY(self.frames.experienceBd, isMounted and self.template.Experience.OffsetY or self.template.Experience.UnmountedOffsetY )
+  end
+end
+
+function GGF.Unit:UpdateShield( value, maxValue )
+  if not self.template.Health.Shield then return end
+  if value == nil then value = 0 end
+  if maxValue == nil then maxValue = 0 end
+  self.shield = {["current"] = value, ["max"] = maxValue, ["percent"] = math.floor( ( value / maxValue ) * 100 )}
+  self.frames.shield:SetWidth( math.min( (value/self.health.max ) , 1 ) * self.template.Health.Bar.Width )
 end
 
 function GGF.Unit:SetName( name )
@@ -138,28 +151,25 @@ end
 function GGF.Unit:SetDifficulty( difficulty )
   if not self.template.Difficulty then return end
   self.difficulty = difficulty
-  -- if self.difficulty == MONSTER_DIFFICULTY_EASY then
-  --   self.frames.difficultyTx:SetTexture('ggFrames\\theme\\themes\\textures\\MonsterDiff1.dds')
-  --   self.frames.difficultyTx:SetHidden(false)
   if self.difficulty == MONSTER_DIFFICULTY_NORMAL then
-    self.frames.difficultyTx:SetTexture('ggFrames\\theme\\themes\\textures\\MonsterDiff1.dds')
+    self.frames.difficultyTx:SetTexture('ggFrames\\theme\\textures\\MonsterDiff1.dds')
     self.frames.difficultyTx:SetHidden(false)
   elseif self.difficulty == MONSTER_DIFFICULTY_HARD then
-    self.frames.difficultyTx:SetTexture('ggFrames\\theme\\themes\\textures\\MonsterDiff2.dds')
+    self.frames.difficultyTx:SetTexture('ggFrames\\theme\\textures\\MonsterDiff2.dds')
     self.frames.difficultyTx:SetHidden(false)
   elseif self.difficulty == MONSTER_DIFFICULTY_DEADLY then
-    self.frames.difficultyTx:SetTexture('ggFrames\\theme\\themes\\textures\\MonsterDiff3.dds')
+    self.frames.difficultyTx:SetTexture('ggFrames\\theme\\textures\\MonsterDiff3.dds')
     self.frames.difficultyTx:SetHidden(false)
   else
     self.frames.difficultyTx:SetHidden(true)
   end
 end
 
-function GGF.Unit:SetClass( className )
+function GGF.Unit:SetClass( classId )
   if not self.template.Class then return end
-  self.class = className
-  if className then
-    self.frames.classTx:SetTexture( GGF.classTextures[className] )
+  self.class = classId
+  if classId then
+    self.frames.classTx:SetTexture( GGF.classTextures[classId] )
     self.frames.classTx:SetHidden(false)
   else
     self.frames.classTx:SetHidden(true)
@@ -167,7 +177,7 @@ function GGF.Unit:SetClass( className )
 end
 
 function GGF.Unit:SetExp( current, max, veteran )
-  if not self.template.Experience then return end
+  if not self.template.Experience or max == 0 then return end
   local exp = current / max
   if self.template.Experience.Label ~= false then
     self.frames.experienceLb:SetText( string.format("%d / %d", current, max) )
@@ -211,12 +221,6 @@ function GGF.Unit:SetRange( isWithinRange )
   self.frames.main:SetAlpha( isWithinRange and 1 or 0.5 )
 end
 
--- GetUnitDifficulty
--- MONSTER_DIFFICULTY_NORMAL
--- MONSTER_DIFFICULTY_EASY
--- MONSTER_DIFFICULTY_HARD
--- MONSTER_DIFFICULTY_DEADLY
--- MONSTER_DIFFICULTY_NONE